+++ packages/appinio_video_player/lib/src/custom_video_player_controller.dart
@@ -23,6 +23,7 @@ class CustomVideoPlayerController {
   VideoPlayerController videoPlayerController;
   final CustomVideoPlayerSettings customVideoPlayerSettings;
   final Map<String, VideoPlayerController>? additionalVideoSources;
+  final ValueNotifier<bool> areControlsVisible = ValueNotifier<bool>(true);
 
   CustomVideoPlayerController({
     required this.context,