+++ packages/appinio_video_player/lib/src/models/custom_video_player_settings.dart
@@ -12,6 +12,12 @@ class CustomVideoPlayerSettings {
   /// Define a custom aspect ratio for the video.
   final double? customAspectRatio;
 
+  /// Define a custom aspect ratio for the video.
+  final Duration seekDuration;
+
+  /// show seek buttons
+  final bool showSeekButtons;
+
   /// The widget that should be shown as long the video is not initialized.
   final Widget? placeholderWidget;
 
@@ -114,6 +120,8 @@ class CustomVideoPlayerSettings {
     this.customAspectRatio,
     this.placeholderWidget,
     this.thumbnailWidget,
+    this.seekDuration = const Duration(seconds: 10),
+    this.showSeekButtons = false,
     this.alwaysShowThumbnailOnVideoPaused = false,
     this.controlsPadding = const EdgeInsets.all(5),
     this.controlBarPadding = const EdgeInsets.all(5),