+++ packages/appinio_video_player/CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.2.2] - 2023.08.23
+
+- Feat: Jump forward and backward like youtube.
+- ```showSeekButtons``` and ```seekDuration``` added to ```CustomVideoPlayerSettings```
+
 ## [1.2.1] - 2023.04.26
 
 - Fix: Orientation not switching back to portrait in iOS